+++ app/src/main/java/com/kuhrusty/z15/model/Scenario.java
@@ -1,5 +1,7 @@
 package com.kuhrusty.z15.model;
 
+import android.support.annotation.NonNull;
+
 public class Scenario {
     public enum Difficulty {
         Prologue, Beginner, Survivor, Hero
@@ -11,6 +13,21 @@ public class Scenario {
     private ZombieDeckShuffler zdShuffler;
     private int cardsPerGrowl = 1;
     private Map map;
+
+    /**
+     * Copies non-null members from the other Scenario into this one.  (This is
+     * so that JSONScenarioRepository can load one file of language-independent
+     * stuff, and another file of language-dependent stuff, and merge the two.)
+     */
+    public void mergeFrom(@NonNull Scenario other) {
+        if ((id == null) && (other.id != null)) id = other.id;
+        if ((name == null) && (other.name != null)) name = other.name;
+        if ((difficulty == null) && (other.difficulty != null)) difficulty = other.difficulty;
+        if ((soundtrack == null) && (other.soundtrack != null)) soundtrack = other.soundtrack;
+        if ((zdShuffler == null) && (other.zdShuffler != null)) zdShuffler = other.zdShuffler;
+        if ((cardsPerGrowl == 1) && (other.cardsPerGrowl != 1)) cardsPerGrowl = other.cardsPerGrowl;
+        if ((map == null) && (other.map != null)) map = other.map;
+    }
 
     public String getID() {
         return id;